+++ source/scan-tool-mso/ChangeLog.md
@@ -1,5 +1,10 @@
 # ChangeLog for scan-tool-mso
 
+## Version 0.04 (2016-03-22)
+  - raise file size limit for scans to 140 MB
+  - use new *.metadefender.com URLs instead of old *.metascan-online.com URLs
+    for scanner requests
+
 ## Version 0.03 (2016-02-24)
   - add "burst mode" which does not honour time limits /rate limits